Add fallback when Worker is not available
