--- src/classes/devices/acaia/common.worker.ts
@@ -1,7 +1,0 @@
-import {DEBUG, WorkerResult, WorkerResultType} from './common';
-
-export function log(...args) {
-  if (DEBUG) {
-    self.postMessage({type: WorkerResultType.LOG, data: args} as WorkerResult);
-  }
-}
