+++ src/classes/devices/acaia/index.ts
@@ -1,0 +1,1 @@
+export {AcaiaScale, EventType} from './acaia';
